Added products component and grouped with edit component in Main

+++ app/Http/Controllers/ProductController.php
@@ -9,11 +9,7 @@ class ProductController extends Controller
 {
     public function index() {
         $products = Product::all();
-        return view('products.index', ['products' => $products]);
-    }
-
-    public function create() {
-        return view('products.create');
+        return response()->json($products);
     }
 
     public function store(Request $request) {
@@ -25,17 +21,51 @@ class ProductController extends Controller
             'description' => 'required'
         ]);
 
-        $newProduct = Product::create($data);
+        Product::create($data);
 
-        return redirect(route('product.index'));
+        return response()->json([
+            "message" => "Product added"
+        ]);
+
     }
 
-    public function edit(Product $product) {
-        return view('products.edit', ['product' => $product]);
+    public function get($id) {
+        $product = Product::find($id);
+        if (!empty($product)) {
+            return response()->json($product);
+        } else {
+            return response()->json([
+                "message" => "Product not found"
+            ], 404);
+        }
     }
 
-    public function delete(Product $product) {
-        $product->delete();
-        return redirect(route('product.index'))->with('success', 'Product deleted succesfully');
+    public function update(Request $request, $id) {
+        if (Product::where('id', $id)->exists()) {
+            $product = Product::find($id);
+            $product->save();
+            return response()->json([
+                "message" => "Product updated"
+            ], 202);
+        } else {
+            return response()->json([
+                "message" => "Product not found"
+            ], 404);
+        }
+    }
+
+    public function delete($id) {
+        if (Product::where('id', $id)->exists()) {
+            $product = Product::find($id);
+            $product->delete();
+
+            return response()->json([
+                "message" => "Product deleted"
+            ], 202);
+        } else {
+            return response()->json([
+                "message" => "Product not found"
+            ], 404);
+        }
     }
 }
